Add script for running FastSurfer preprocessing
This commit adds a new script, srun_fastsurfer_user.sh, which is used for running the preprocessing step of FastSurfer. The script sets up the necessary SLURM job parameters and defines the paths for input data, output files, and the FastSurfer image. It also sets the temporary directory and executes the FastSurfer preprocessing command using Singularity.

The script is currently commented out, as the Singularity image and license arguments are not provided. The commented out command can be uncommented and modified with the appropriate paths and arguments.

Note: The script does not have a newline at the end of the file.